refactor(knowledge): extract searchKnowledgeBase into core service to break core->tools cycle
